% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binarray}
\alias{binarray}
\title{Bin arrays into blocks}
\usage{
binarray(x, blocksize = c(2, 2), reshapeTo2D = FALSE)
}
\arguments{
\item{x}{matrix/array: The matrix to divide into blocks}

\item{blocksize}{vector: dimensions of the blocks}

\item{reshapeTo2D}{Bool: Flag to set to TRUE if instead of wanting an output array of dimension nxmxk, you want the output to be dimension (nm)xk, i.e. you want to flatten the blocks. Default FALSE}
}
\value{
array blocked into chunks of the desired size
}
\description{
Takes a 2D array and divides it into blocks of constant size. Probably not the fastest possible implementation, but for now this is what we are doing.
}
