% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPS.R
\name{CRPS}
\alias{CRPS}
\title{Calculate the CRPS for an arbitrary univariate distribution}
\usage{
CRPS(x, pdf = NULL, cdf = NULL, xhat = NULL)
}
\arguments{
\item{x}{vector. the values at which the PDF or CDF are evaluated}

\item{pdf}{vector. the predictive PDF evaluated at the values in x. Default NULL. One of PDF or CDF must not be NULL}

\item{cdf}{vector. the predictive CDF evaluated at the values in x. Default NULL. One of PDF or CDF must not be NULL}

\item{xhat}{scalar. The observed value to compare with the predictive distribution.}
}
\value{
The continuous rank probability score for the predictive model given the data, calculated via the integral 
$\int_{R}(F(y)-H(y-xhat))^2 dy$.
}
\description{
Calculates the CRPS of an arbitrary predictive distribution given the predictive PDF or CDF and the observed value
}
\details{
Functions for calculating the Continuous Rank Probability Score for predictive models
}
