% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPS.R
\name{CRPSgaussian}
\alias{CRPSgaussian}
\title{Calculate the CRPS for a predictive Gaussian distribution}
\usage{
CRPSgaussian(mu, sigma, xhat)
}
\arguments{
\item{mu}{scalar. the mean of the predictive distribution}

\item{sigma}{scalar. the standard deviation of the predictive distribution}

\item{xhat}{scalar. The observed value to compare with the predictive distribution.}
}
\value{
The continuous rank probability score for the predictive model given the data, using the closed form for the univariate Gaussian from "Strictly Proper Scoring Rules, Prediction, and Estimation." by Tilmann Gneiting &Adrian E Raftery
}
\description{
Calculate the CRPS for a predictive Gaussian distribution with parameters mu and sigma, and observed value xhat using the closed form presented in
"Strictly Proper Scoring Rules, Prediction, and Estimation." by Tilmann Gneiting &Adrian E Raftery
}
