% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPS.R
\name{CRPSgbp}
\alias{CRPSgbp}
\title{Generalized Beta prime CRPS}
\usage{
CRPSgbp(alpha, beta, xhat, p = 1, q = 1)
}
\arguments{
\item{alpha}{Shape parameter (alpha > 0)}

\item{beta}{Shape parameter (beta > 1/p for finite mean)}

\item{p}{Shape parameter (p > 0)}

\item{q}{Scale parameter (q > 0)}

\item{y}{Observation (scalar, positive)}
}
\value{
CRPS value (scalar)
}
\description{
Compute the CRPS of a Generalized Beta Prime distribution
The GBP pdf is:
  f(z) = p/(q*B(alpha,beta)) * (z/q)^(alpha*p-1) / (1+(z/q)^p)^(alpha+beta)
Reference: Eq. (15) of LeDuc 2026 The Continuous Rank Probability Score of a Generalized
Beta-Prime Distribution and Some Special Cases
Due to issues with evaluating generalized hypergeometric functions at unit argument using
the hypergeo package, this method computes \eqn{\mathbb{E}[XF(X)]} using quadrature.
}
