% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaPrimeFunctions.R
\name{dbetaprime}
\alias{dbetaprime}
\title{Beta-prime PDF}
\usage{
dbetaprime(x, alphaparam, betaparam, p = 1, q = 1)
}
\arguments{
\item{x}{scalar or matrix. Values at which to evaluate the PDF}

\item{alphaparam}{Scalar value, alpha parameter of the distribution}

\item{betaparam}{Scalar value, beta parameter of the distribution}

\item{p}{Scalar value, p parameter of the distribution}

\item{q}{Scalar value, q parameter of the distribution}
}
\value{
The beta-prime pdf with input parameters evaluated at the locations in x. Check values, possible issues for very large alphaval or betaval parameters
}
\description{
Evaluates the Beta-Prime PDF with parameters alpha,beta,p,q (parameterization used in https://en.wikipedia.org/wiki/Beta_prime_distribution )
}
\details{
Functions for doing basic calculations with the beta-prime distribution
}
