% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPS.R
\name{e_2xfx_gq}
\alias{e_2xfx_gq}
\title{Numerical evaluation of \eqn{\mathbb{E}[2XF(X)]} for the generalized Beta-prime dsitribution}
\usage{
e_2xfx_gq(alpha, beta, p = 1, q = 1, quadrule = NULL, n = 64)
}
\arguments{
\item{alpha}{scalar, the alpha parameter}

\item{beta}{scale, the beta parameter}

\item{p}{the p parameter, default 1}

\item{q}{the q parameter, default 1}

\item{quadrule}{a quadrature rule on [0,1] generated by statmod::gauss.quad. Default NULL, in which case it is generated in the function}

\item{n}{the number of nodes on the quadrature rule, default is \eqn{ceil(min(64, \sqrt(\alpha+\beta))}}
}
\description{
Uses a quadrature rule on [0,1] to compute the value of the expectation
\eqn{\mathbb{E}[2XF(X)]} for use in calculating the CRPS of the generalized Beta-prime. This is done
because the package hypergeo() can be poorly behaved when computing generalized
hypergeometric functions at argument 1
}
