% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gausssmooth}
\alias{gausssmooth}
\title{Gaussian smoother}
\usage{
gausssmooth(data, sigma)
}
\arguments{
\item{sigma}{scalar: Width of the Gaussian smoother}
}
\value{
Smoothed data
}
\description{
Applies a Gaussian smoother with width sigma to the input data
}
