% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getsza}
\alias{getsza}
\title{Solar zenith angle}
\usage{
getsza(lat, lon, hourUTC, julianDay)
}
\arguments{
\item{lat}{Scalar or matrix. Latitude of the points at which to calculate SZA. 0=equator, Degrees N.}

\item{lon}{Scalar or matrix. Longitude of the points at which to calculate SZA. Degrees E, ranging from -180 to 180}

\item{hourUTC}{Scalar. Time in UTC, hours since midnight}

\item{julianDay}{Scalar. Julian day, AKA ordinal day, i.e. 1 = Jan1, 2 = Jan 2, 32 = Feb 1, etc.}
}
\value{
Solar zenith angle, in degrees
}
\description{
Calculates the solar zenith angle at a particular point using the latitude, longitude, time in UTC, and julian day
}
\details{
Utility functions for the package PoissonRatioUQ
}
