% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermanentalProcessEstimation.R
\name{gradientFn}
\alias{gradientFn}
\title{Gradient of the RKHS-regularized Poisson point process likelihood with respect to the kernel coefficients}
\usage{
gradientFn(alpha, counts, K)
}
\arguments{
\item{alpha}{vector: Kernel coefficients \eqn{\vec{\alpha}} so that \eqn{f(x)= \sum \alpha_i \tilde{k}(x_i,x)}}

\item{counts}{vector: Count data, number observed at each observation point}

\item{K}{matrix: Kernel matrix for the equivalent kernel function \eqn{\tilde{k}(x,y)}}
}
\value{
the gradient of the objective function at alpha
}
\description{
Gradient of the RKHS-regularized Poisson point process likelihood with respect to the kernel coefficients
}
