% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDSet.R
\name{hpdintervalgaussian}
\alias{hpdintervalgaussian}
\title{Highest-Posterior density for a Gaussian distribution}
\usage{
hpdintervalgaussian(mu, sigma, alpha)
}
\arguments{
\item{mu}{Scalar. Mean of the Gaussian distribution}

\item{sigma}{scalar. Standard deviation of the Gaussian distribution}

\item{alpha}{scalar. The desired credibility value, or the desired value of the integral of the pdf over the HPD set.}
}
\value{
The endpoints of the alpha-HPD interval.
}
\description{
Calculates the highest-posterior density set of the univariate Gaussian with mean mu and standard deviation sigma
}
