% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDSet.R
\name{hpdset}
\alias{hpdset}
\title{Highest-Posterior density}
\usage{
hpdset(x, fx, alpha)
}
\arguments{
\item{x}{vector. x value of x at which the PDF is evaluated}

\item{fx}{vector. Values of the PDF at the points in x}

\item{alpha}{scalar. The desired credibility value, or the desired value of the integral of the pdf over the HPD set.}
}
\value{
The values of x within the alpha-highest posterior density set
}
\description{
Calculates the highest-posterior density set of an abirtrary univariate distribution given a PDF fx evaluated at points x
}
\details{
Functions for computing the highest posterior density set of a given distribution
}
