% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPS.R
\name{inc_beta}
\alias{inc_beta}
\title{Incomplete Beta function}
\usage{
inc_beta(w, a, b)
}
\arguments{
\item{w}{Locations to evaluate}

\item{a}{the a parameter}

\item{b}{the b parameter}
}
\value{
the value of the incomplete beta function at w
}
\description{
Evaluates the incomplete beta function at nodes w with parameters a,b
}
