% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermanentalProcessEstimation.R
\name{objectivefn}
\alias{objectivefn}
\title{Objective function of the RKHS-regularized Poisson point process likelihood with respect to the kernel coefficients}
\usage{
objectivefn(alpha, counts, K, c)
}
\arguments{
\item{alpha}{vector: Kernel coefficients alpha so that \eqn{\vec{f}=\tilde{K}\vec{\psi}}}

\item{counts}{vector: Count data, number observed at each observation point}

\item{K}{matrix: Kernel matrix for the kernel function \eqn{\tilde{k}}}

\item{c}{scalar: Value so that \eqn{\vec{\Lambda} = \frac{c}{2}\vec{f}^2}}
}
\value{
the value of the objective function at alpha
}
\description{
Objective function of the RKHS-regularized Poisson point process likelihood with respect to the kernel coefficients
}
