% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaPrimeFunctions.R
\name{pbetaprime}
\alias{pbetaprime}
\title{Beta-prime CDF}
\usage{
pbetaprime(x, alphaparam, betaparam, p = 1, q = 1)
}
\arguments{
\item{x}{scalar or matrix. Values at which to evaluate the CDF}

\item{alphaparam}{Scalar value, alpha parameter of the distribution}

\item{betaparam}{Scalar value, beta parameter of the distribution}

\item{p}{Scalar value, p parameter of the distribution}

\item{q}{Scalar value, q parameter of the distribution}
}
\value{
The beta-prime cdf with input parameters evaluated at the locations in x. Check values, possible issues for very large alphaval or betaval parameters
}
\description{
Evaluates the Beta-Prime CDF with parameters alpha,beta,p,q (parameterization used in https://en.wikipedia.org/wiki/Beta_prime_distribution#Generalization )
}
