% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermanentalProcessEstimation.R
\name{permprocest}
\alias{permprocest}
\title{Estimation of point process intensities using a Permanental Process model}
\usage{
permprocest(K, counts = NaN, g = 1, c = 1, maxiter = 300)
}
\arguments{
\item{K}{matrix: Kernel matrix evaluated at the spatial locations}

\item{counts}{numeric: The number of counts in each bin.}

\item{g}{scalar: Regularization strength (or prior precision). Default is 1}

\item{c}{scalar: Parameter such that intensity \eqn{\vec{\Lambda} = \frac{c}{2}\vec{f}^2}, default is 1.}

\item{maxiter}{scalar: The maximum number of iterations.}
}
\value{
The estimated intensities, the parameters of the gamma approximation to its posterior at each location, and the fitted kernel coefficients of \eqn{\hat{f} = \tilde{K}\vec{\psi}}
}
\description{
Given Possion data, estimates the underlying intensity by assuming that it has the form \eqn{\vec{\Lambda} = \frac{c}{2}\vec{f}^2}, with \eqn{\vec{f}} a Gaussian vector. The algorithm recovers a latent vector \eqn{\vec{f}} under the prior \eqn{\vec{f}\sim\mathcal{N}(0,g^{-1}K)}, and the result is given by \eqn{\vec{f}=\tilde{K}\vec{\psi}} for some \eqn{\vec{\psi}} to be retrieved. The algorithm is adapted from the one outlined in https://proceedings.mlr.press/v70/walder17a.html and was used in https://egusphere.copernicus.org/preprints/2025/egusphere-2025-5570/ to estimate thermospheric temperature
}
\details{
Functions for doing estimation of the parameters of a permanental process model using equivalent kernel representations
}
