% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaPrimeFunctions.R
\name{qbetaprime}
\alias{qbetaprime}
\title{generalized Beta-prime quantile function}
\usage{
qbetaprime(probs, alphaparam, betaparam, p = 1, q = 1)
}
\arguments{
\item{probs}{scalar or matrix. Probabilities at which to evaluate the quantile function.}

\item{alphaparam}{Scalar value, alpha parameter of the distribution}

\item{betaparam}{Scalar value, beta parameter of the distribution}

\item{p}{Scalar value, p parameter of the distribution}

\item{q}{Scalar value, q parameter of the distribution}
}
\value{
The beta-prime quantile function with input parameters evaluated at the probabilities. Check values, possible issues for very large alphaparam or betaparam parameters
}
\description{
Evaluates the generalized Beta-Prime quantile function with parameters alpha,beta,p,q (parameterization used in https://en.wikipedia.org/wiki/Beta_prime_distribution#Generalization ).This uses the fact that if X ~ BP(a,b,p,q) then (X/q)^p  ~ BP(a,b)
}
