% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermanentalProcessEstimation.R
\name{ratioestimationpermproc}
\alias{ratioestimationpermproc}
\title{Estimation of the ratio of point process intensities using a Permanental Process model}
\usage{
ratioestimationpermproc(
  K1,
  counts1,
  counts2,
  K2 = K1,
  c1 = 1,
  g1 = 1,
  c2 = c1,
  g2 = g1,
  maxiter = 300
)
}
\arguments{
\item{K1}{matrix: Kernel matrix for the numerator evaluated at the data points.}

\item{counts1}{numeric: Count data for the numerator, see documentation for permprocest(). Must be the same type of process as the denominator process, i.e. both binned or both unbinned.}

\item{counts2}{numeric: Count data for the denominator, see documentation for permprocest(). Must be the same type of process as the numerator process, i.e. both binned or both unbinned.}

\item{K2}{matrix: Kernel matrix for the denominator evaluated at the data points. Default is K1.}

\item{c1}{scalar: Parameter such that intensity \eqn{\lambda(s) = \frac{c}{2}f(s)^2} for the problem in the numerator. Default is 1.}

\item{g1}{scalar: Regularization strength (or prior precision) for the problem in the numerator. Default is 1}

\item{c2}{scalar: Parameter such that intensity \eqn{\lambda(s) = \frac{c}{2}f(s)^2} for the problem in the denominator. Default is c1.}

\item{g2}{scalar: Regularization strength (or prior precision) for the problem in the denominator. Default is g1}

\item{maxiter}{scalar: The maximum number of iterations.}
}
\value{
The estimated intensity ratio (MAP) and the parameters of the Generalized Beta-Prime distribution describing the posterior
}
\description{
Given a realizataion of two point processes, estimates the ratios of the underlying intensities by assuming that each intensity function is a squared Gaussian process.
}
