% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowVars}
\alias{rowVars}
\title{Row variances}
\usage{
rowVars(x, na.rm = TRUE)
}
\arguments{
\item{x}{matrix: The matrix}

\item{na.rm}{BOOL: Whether of not you want NaNs removed. Default is TRUE}
}
\value{
Solar zenith angle, in degrees
}
\description{
Calculates the variance of each row of a matrix
}
