% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributions.R
\name{tgivenab}
\alias{tgivenab}
\title{Determine the parameters of the distribution of the temperature given the data under the assumption that Z is either a point estimate or has a Gaussian distribution, and is an affine function of T}
\usage{
tgivenab(a, b, m, z0, tausq, mu0 = 0, sigma0 = Inf, uncertainty = "Gaussian")
}
\arguments{
\item{a}{Matrix. The count data for the numerator. Rows correspond to spatial locations, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. Missing data should be replaced by NaNs.}

\item{b}{Matrix. The count data for the denominator. Rows correspond to spatial locations, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. Missing data should be replaced by NaNs.}

\item{m}{scalar. slope parameter of the linear regression of Z(T). Must be a scalar for now, will be generalized to a vector in the future.}

\item{z0}{scalar. Intercept parameter of the linear regression of Z(T). Must be a scalar for now, will be generalized to a vector in the future.}

\item{tausq}{scalar. Variance of the residuals of the linear regression. Must be a scalar for now, will be generalized to a vector in the future.}

\item{mu0}{scalar. Mean of Gaussian prior for the temperature. Default is 0. Must be a scalar for now, will be generalized to a vector in the future.}

\item{sigma0}{scalar. Standard deviation of the Gaussian prior for the temperature. Default is Inf (uninformative prior). Must be a scalar for now, will be generalized to a vector in the future.}

\item{uncertainty}{string. Either "Gaussian", in which case Z is assumed to have a Gaussian distribution determined by the function zgaussian, or "none", in which case mean(a)/mean(b) is used as a point estimate of Z. Default is "Gaussian"}
}
\value{
mean and standard deviation of the distribution T|a,b with Z=mT+z0 and T ~ N(mu0,sigma0^2). In the future when more spatial structure is allowed this will return a mean and covariance matrix.
}
\description{
Determine the parameters of the distribution of the temperature given the data under the assumption that Z is either a point estimate or has a Gaussian distribution, and is an affine function of T
}
