% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributions.R
\name{zgaussian}
\alias{zgaussian}
\title{Calculate the distribution of Z under a Gaussian assumption}
\usage{
zgaussian(a, b)
}
\arguments{
\item{a}{vector. The count data for the numerator}

\item{b}{vector. The count data for the denominator}
}
\value{
a list containing the mean and standard deviation of the approrpiate Gaussian
}
\description{
given count datasets a and b, calculates the Gaussian approximation of the distribution of Z = \lambda_a / \lambda_b where \lambda_i is the Poisson intensity of the channel that produced dataset i
}
\details{
Functions for calculating distributions given data
}
