% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributions.R
\name{tgivenab}
\alias{tgivenab}
\title{Determine the parameters of the distribution of the temperature given the data under the assumption that Z is either a point estimate or has a Gaussian distribution, and is an affine function of T}
\usage{
tgivenab(a, b, m, z0, tausq, mu0 = 0, sigma0 = Inf, uncertainty = "Gaussian")
}
\arguments{
\item{a}{vector. The count data for the numerator}

\item{b}{vector. The count data for the denominator}

\item{m}{scalar. slope parameter of the linear regression of Z(T)}

\item{z0}{scalar. Intercept parameter of the linear regression of Z(T)}

\item{tausq}{scalar. Variance of the residuals of the linear regression}

\item{mu0}{scalar. Mean of Gaussian prior for the temperature. Default is 0.}

\item{sigma0}{scalar. Standard deviation of the Gaussian prior for the temperature. Default is Inf (uninformative prior)}

\item{uncertainty}{string. Either "Gaussian", in which case Z is assumed to have a Gaussian distribution determined by the function zgaussian, or "none", in which case mean(a)/mean(b) is used as a point estimate of Z. Default is "Gaussian"}
}
\value{
mean and standard deviation of the distribution T|a,b with Z=mT+z0 and T ~ N(mu0,sigma0^2)
}
\description{
Determine the parameters of the distribution of the temperature given the data under the assumption that Z is either a point estimate or has a Gaussian distribution, and is an affine function of T
}
