% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributions.R
\name{tgivenab}
\alias{tgivenab}
\title{Determine the parameters of the distribution of the temperature given the data under the assumption that Z has a Beta-Prime distribution and that \eqn{Z=(mT+z_0)^p}. This leads to T having a shifted Beta-Prime distribution.}
\usage{
tgivenab(
  a,
  b,
  m,
  z0,
  p,
  spatial = TRUE,
  a1 = 1,
  a2 = 1,
  b1 = 0,
  b2 = 0,
  K1 = NA,
  K2 = K1,
  g1 = 1,
  c1 = 1,
  g2 = g1,
  c2 = c1
)
}
\arguments{
\item{a}{Matrix. The count data for the numerator. Rows correspond to spatial locations/channels, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. MUST be a matrix or array right now, if you want a scalar use a 1x1 array. Missing data should be replaced by NaNs.}

\item{b}{Matrix. The count data for the denominator. Rows correspond to spatial locations/channels, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. MUST be a matrix or array right now, if you want a scalar use a 1x1 array. Missing data should be replaced by NaNs.}

\item{m}{scalar or vector. Slope parameter of the model Z(T)=(mT+z_0)^p.}

\item{z0}{scalar or vector. Intercept parameter of the model Z(T)=(mT+z_0)^p.}

\item{p}{scalar or vector. Exponent parameter of the model Z(T)=(mT+z_0)^p.}

\item{spatial}{boolean. If TRUE (default), run the permanental process model. If FALSE, runs the pointwise estimation scheme.}

\item{a1, a2, b1, b2}{scalars: Parameters of the prior distribution for the pointwise model. See documentation for zbetaprime()}

\item{K1, K2}{Matrices. Kernel matrices for the permanental process model. See ratioestimationpermproc() for details.}

\item{g1, g2, c1, c2}{scalars, other parameters for the permanental process model. See ratioestimationpermproc() for more details. This funciton assumes they are the same for both models.}
}
\value{
Parameters of the shifted generalized Beta-Prime distribution T|a,b using the permanental process model, where the band ratio Z(T)=(mT+z_0)^p.
}
\description{
Determines the distribution of T|data under the assumption that the counts are Poisson distributed, or a realization of a Poisson Point Process. Can perform pointwise estimation or estimation using a permanental process model that allows for explicit modeling of spatial correlation.
}
\details{
Functions for calculating distributions given data
}
