% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributions.R
\name{tgivenab}
\alias{tgivenab}
\title{Determine the parameters of the distribution of the temperature given the data under the assumption that Z is either a point estimate or has a Gaussian distribution, and is an affine function of T}
\usage{
tgivenab(
  a,
  b,
  M,
  Z0,
  TauSq,
  priormn = 0,
  priorvar = Inf,
  uncertainty = "Gaussian",
  spectralAveraging = FALSE
)
}
\arguments{
\item{a}{Matrix. The count data for the numerator. Rows correspond to spatial locations, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. MUST be a matrix or array right now, if you want a scalar use a 1x1 array. Missing data should be replaced by NaNs.}

\item{b}{Matrix. The count data for the denominator. Rows correspond to spatial locations, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. MUST be a matrix or array right now, if you want a scalar use a 1x1 array. Missing data should be replaced by NaNs.}

\item{M}{scalar or matrix. slope parameter of the linear regression of Z(T).}

\item{Z0}{scalar or vector. Intercept parameter of the linear regression of Z(T).}

\item{TauSq}{scalar or matrix. Covariance matrix of the residuals of the linear regression.}

\item{priormn}{scalar or vector. Mean of Gaussian prior for the temperature. Default is 0.}

\item{priorvar}{scalar or matrix. variance of the Gaussian prior for the temperature. Default is Inf (uninformative prior).}

\item{uncertainty}{string. Either "Gaussian", in which case Z is assumed to have a Gaussian distribution determined by the function zgaussian, or "none", in which case mean(a)/mean(b) is used as a point estimate of Z. Default is "Gaussian"}
}
\value{
mean and standard deviation of the distribution T|a,b with Z=mT+z0, T|Z ~ N((a/b-z0)/m, tau^2/m), Z|a,b from zgaussian(), and T ~ N(mu0,Sigma0) where mu0 is the prior mean and Sigma0 the prior covariance matrix.
}
\description{
Given count data, a prior, and the parameters of the distribution Z|T ~N(mT+z0,tausq) determines the distribution of T|data under the assumption that the counts are either Gaussian or provide a point estimate of Z.
}
