% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributions.R
\name{zbetaprime}
\alias{zbetaprime}
\title{Calculate the distribution of Z under a Beta-Prime assumption}
\usage{
zbetaprime(a, b, k1 = 0, k2 = 0)
}
\arguments{
\item{a}{vector. The count data for the numerator}

\item{b}{vector. The count data for the denominator}

\item{k1}{scalar. The prior for the intensity of the channel generating the dataset {a}, of the form x^{-k_1}. k_1=0 is an uninformative prior, k_1=1 is scale-invariant}

\item{k2}{scalar. The prior for the intensity of the channel generating the dataset {b}, of the form x^{-k_2}. k_2=0 is an uninformative prior, k_2=1 is scale-invariant}
}
\value{
a list containing the parameters of the beta-prime distribution, using the parameterization in https://en.wikipedia.org/wiki/Beta_prime_distribution
}
\description{
given count datasets a and b, calculates the distribution of Z = \lambda_a / \lambda_b where \lambda_i is the Poisson intensity of the channel that produced dataset i. Under a Gamma prior for each intensity, this is a Beta-Prime distribution
}
