% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributions.R
\name{zgaussian}
\alias{zgaussian}
\title{Calculate the distribution of Z under a Gaussian assumption}
\usage{
zgaussian(a, b)
}
\arguments{
\item{a}{Matrix. The count data for the numerator. Rows correspond to spatial locations, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. Missing data should be replaced by NaNs.}

\item{b}{Matrix. The count data for the denominator. Rows correspond to spatial locations, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. Missing data should be replaced by NaNs.}
}
\value{
a list containing the means and standard deviations of the appropriate Gaussians.
}
\description{
given count datasets a and b, calculates the Gaussian approximation of the distribution of Z = lambda_a / lambda_b where lambda_i is the Poisson intensity of the channel that produced dataset i. Some details are available in (Park, T., Kashyap, V. L., Siemiginowska, A., van Dyk, D. A., Zezas, A., Heinke, C., and Wargelin, B. J.: Bayesian Estimation of Hardness
Ratios: Modeling and Computations, In: The
Astrophysical Journal, 2006) and (Gehrels, N.: Confidence limits for small numbers of events in astrophysical data, The Astrophysical Journal, 1986.). This model assumes that realizations at different spatial locations are independent!
}
\details{
Functions for calculating distributions given data
}
