% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistributions.R
\name{zgaussian}
\alias{zgaussian}
\title{Calculate the distribution of Z under a Gaussian assumption}
\usage{
zgaussian(a, b, a_prior = c(1, 0), b_prior = c(1, 0))
}
\arguments{
\item{a}{Matrix. The count data for the numerator. Rows correspond to spatial locations, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. MUST be a matrix or array right now, if you want a scalar use a 1x1 array. Missing data should be replaced by NaNs.}

\item{b}{Matrix. The count data for the denominator. Rows correspond to spatial locations, columns to realizations. Must have the same number of rows as b, but need not have the same number of columns. MUST be a matrix or array right now, if you want a scalar use a 1x1 array. Missing data should be replaced by NaNs.}

\item{a_prior}{list/array. The shape and rate parameters of the Gamma prior distribution for the Poisson intensity of the numerator process. Default is (1,0), i.e. uniform prior.}

\item{b_prior}{list/array. The shape and rate parameters of the Gamma prior distribution for the Poisson intensity of the denominator process. Default is (1,0), i.e. uniform prior.}
}
\value{
a list containing the means and standard deviations of the appropriate Gaussians.
}
\description{
given count datasets a and b, calculates the Gaussian approximation of the distribution of Z = lambda_a / lambda_b, where lambda_i is the Poisson intensity of the channel that produced dataset i, using the delta method. Assumes a Gamma prior on Poisson data to get the distribution of the intensities, and then uses the delta method to get the approximate distribution of the ratio.
This model assumes that realizations at different spatial locations are independent.
}
\details{
Functions for calculating distributions given data
}
